refactoring locator
